feat(windows): use double precision

+++ packages/audioplayers_windows/windows/audio_player.h
@@ -69,9 +69,9 @@ class AudioPlayer {
 
     bool GetLooping();
 
-    int64_t GetPosition();
+    double GetPosition();
 
-    int64_t GetDuration();
+    double GetDuration();
 
     void SeekTo(double seek);
